Extract image generation/cache from texturesource.cpp
